requirements.txt
pandas
realtime-crypto
alpaca-py
numpy
scikit-learn
lightgbm
joblib
matplotlib
pytest